expose freesurfer lookup table
