+++ mne/utils/config.py
@@ -17,7 +17,7 @@ import re
 
 import numpy as np
 
-from .check import _validate_type, _check_pyqt5_version
+from .check import _validate_type, _check_pyqt5_version, _check_fname
 from ._logging import warn, logger
 
 
@@ -366,6 +366,53 @@ def get_subjects_dir(subjects_dir=None, raise_error=False):
     if subjects_dir is not None:
         subjects_dir = str(subjects_dir)
     return subjects_dir
+
+
+def read_freesurfer_lut(fname=None):
+    """Read a Freesurfer-formatted LUT.
+
+    Parameters
+    ----------
+    fname : str | None
+        The filename. Can be None to read the standard Freesurfer LUT.
+
+    Returns
+    -------
+    atlas_ids : dict
+        Mapping from label names to IDs.
+    colors : dict
+        Mapping from label names to colors.
+    """
+    lut = _get_lut(fname)
+    names, ids = lut['name'], lut['id']
+    colors = np.array([lut['R'], lut['G'], lut['B'], lut['A']], float).T
+    atlas_ids = dict(zip(names, ids))
+    colors = dict(zip(names, colors))
+    return atlas_ids, colors
+
+
+def _get_lut(fname=None):
+    """Get a FreeSurfer LUT."""
+    _validate_type(fname, ('path-like', None), 'fname')
+    if fname is None:
+        fname = op.join(op.dirname(__file__), 'data', 'FreeSurferColorLUT.txt')
+    _check_fname(fname, 'read', must_exist=True)
+    dtype = [('id', '<i8'), ('name', 'U'),
+             ('R', '<i8'), ('G', '<i8'), ('B', '<i8'), ('A', '<i8')]
+    lut = {d[0]: list() for d in dtype}
+    with open(fname, 'r') as fid:
+        for line in fid:
+            line = line.strip()
+            if line.startswith('#') or not line:
+                continue
+            line = line.split()
+            if len(line) != len(dtype):
+                raise RuntimeError(f'LUT is improperly formatted: {fname}')
+            for d, part in zip(dtype, line):
+                lut[d[0]].append(part)
+    lut = {d[0]: np.array(lut[d[0]], dtype=d[1]) for d in dtype}
+    assert len(lut['name']) > 0
+    return lut
 
 
 def _get_stim_channel(stim_channel, info, raise_error=True):
